Organize constant data in the same file called constants
