+++ app/src/lib/constants.ts
@@ -1,0 +1,316 @@
+export const body_parts = [
+	'empty',
+	'head',
+	'chest',
+	'upper_abdomen',
+	'left_arm',
+	'right_arm',
+	'lower_abdomen',
+	'left_leg',
+	'right_leg',
+	'left_knee',
+	'right_knee',
+	'left_calves',
+	'right_calves',
+	'left_foot',
+	'right_foot'
+];
+
+export const bodyPartChinese = {
+	head: '頭部',
+	chest: '胸部',
+	left_arm: '左手',
+	right_arm: '右手',
+	upper_abdomen: '上腹部',
+	lower_abdomen: '下腹部',
+	left_knee: '左膝',
+	right_knee: '右膝',
+	left_leg: '左大腿',
+	right_leg: '右大腿',
+	left_calves: '左小腿',
+	right_calves: '右小腿',
+	left_foot: '左腳',
+	right_foot: '右腳'
+};
+
+export const symptomsList = {
+	head: [
+		'頭暈',
+		'頭痛',
+		'幻覺 (幻視、幻聽、幻觸等)',
+		'情緒低落/憂鬱',
+		'失眠',
+		'嗜睡',
+		'記憶衰退/認知障礙',
+		'注意力不集中',
+		'自殘行為或傾向 (含自殺)',
+		'焦慮',
+		'莫名恐慌',
+		'無法控制的食慾變化（厭食、暴食）',
+		'躁動易怒',
+		'言語混亂',
+		'掉髮',
+		'眩暈',
+		'喉嚨痛',
+		'喉嚨異物感',
+		'吞嚥困難',
+		'發音障礙(沙啞、無法出聲、小聲、構音障礙)',
+		'痰多',
+		'頸部腫塊',
+		'噁心/嘔吐',
+		'耳痛',
+		'聽力下降',
+		'耳鳴',
+		'耳聾/聽力損傷',
+		'鼻血',
+		'鼻塞',
+		'鼻涕倒流',
+		'嗅覺下降',
+		'鼻痛',
+		'異味感',
+		'鼻乾',
+		'打鼾',
+		'吞嚥困難',
+		'痰多',
+		'生長發育遲緩',
+		'痘痘',
+		'皮膚搔癢',
+		'皮膚紅腫',
+		'皮膚痛',
+		'皮膚麻木',
+		'水泡 / 囊泡',
+		'發疹',
+		'白斑',
+		'化膿',
+		'脫屑'
+	],
+	chest: ['胸悶', '胸痛', '易喘(依NYHA分級)', '心絞痛', '心悸', '豐胸', '乳汁分泌問題', '乳房脹痛'],
+	left_arm: [
+		'感覺異常',
+		'肢體麻痺/無力',
+		'擦傷',
+		'挫傷',
+		'扭拉傷',
+		'痠、痛',
+		'痠軟無力',
+		'僵硬',
+		'撕裂痛',
+		'活動度受限',
+		'活動時疼痛'
+	],
+	right_arm: [
+		'感覺異常',
+		'肢體麻痺/無力',
+		'擦傷',
+		'挫傷',
+		'扭拉傷',
+		'痠、痛',
+		'痠軟無力',
+		'僵硬',
+		'撕裂痛',
+		'活動度受限',
+		'活動時疼痛'
+	],
+	upper_abdomen: [
+		'噁心/嘔吐',
+		'胃口改變(食慾變大或變小)',
+		'噯腐吞酸/泛酸',
+		'打嗝',
+		'消化不良',
+		'脹氣',
+		'腹痛',
+		'胃灼熱感'
+	],
+	lower_abdomen: [
+		'大小便失禁',
+		'頻尿、尿急',
+		'遺尿、尿不乾淨感',
+		'漏尿、尿失禁',
+		'夜尿',
+		'解尿疼痛、排尿灼熱',
+		'排尿困難、尿瀦留',
+		'茶色尿',
+		'血尿',
+		'泡沫尿',
+		'會陰部疼痛、睪丸、陰莖疼痛',
+		'睪丸、陰莖紅腫/有滲出物',
+		'陰莖、睪丸長出丘疹、不明突起、潰爛等',
+		'生殖器官、陰部腫痛or濕疹',
+		'陽痿、勃起/射精功能問題',
+		'不孕',
+		'性交疼痛',
+		'壯陽',
+		'月經異常',
+		'經痛',
+		'白帶異常',
+		'便祕',
+		'解便疼痛',
+		'肛門紅腫/癢/痛',
+		'血便/黑便/脂肪便',
+		'養胎/懷孕階段調理'
+	],
+	left_leg: [
+		'感覺異常',
+		'肢體麻痺/無力',
+		'擦傷',
+		'挫傷',
+		'扭拉傷',
+		'痠、痛',
+		'痠軟無力',
+		'僵硬',
+		'撕裂痛',
+		'活動度受限',
+		'活動時疼痛',
+		'痘痘',
+		'皮膚搔癢',
+		'皮膚紅腫',
+		'皮膚痛',
+		'皮膚麻木',
+		'水泡 / 囊泡',
+		'發疹',
+		'白斑',
+		'化膿',
+		'脫屑'
+	],
+	right_leg: [
+		'感覺異常',
+		'肢體麻痺/無力',
+		'擦傷',
+		'挫傷',
+		'扭拉傷',
+		'痠、痛',
+		'痠軟無力',
+		'僵硬',
+		'撕裂痛',
+		'活動度受限',
+		'活動時疼痛',
+		'痘痘',
+		'皮膚搔癢',
+		'皮膚紅腫',
+		'皮膚痛',
+		'皮膚麻木',
+		'水泡 / 囊泡',
+		'發疹',
+		'白斑',
+		'化膿',
+		'脫屑'
+	],
+	left_knee: [
+		'感覺異常',
+		'肢體麻痺/無力',
+		'擦傷',
+		'挫傷',
+		'扭拉傷',
+		'痠、痛',
+		'痠軟無力',
+		'僵硬',
+		'撕裂痛',
+		'活動度受限',
+		'活動時疼痛',
+		'痘痘',
+		'皮膚搔癢',
+		'皮膚紅腫',
+		'皮膚痛',
+		'皮膚麻木',
+		'水泡 / 囊泡',
+		'發疹',
+		'白斑',
+		'化膿',
+		'脫屑'
+	],
+	right_knee: [
+		'感覺異常',
+		'肢體麻痺/無力',
+		'擦傷',
+		'挫傷',
+		'扭拉傷',
+		'痠、痛',
+		'痠軟無力',
+		'僵硬',
+		'撕裂痛',
+		'活動度受限',
+		'活動時疼痛',
+		'痘痘',
+		'皮膚搔癢',
+		'皮膚紅腫',
+		'皮膚痛',
+		'皮膚麻木',
+		'水泡 / 囊泡',
+		'發疹',
+		'白斑',
+		'化膿',
+		'脫屑'
+	],
+	left_calves: [
+		'感覺異常',
+		'肢體麻痺/無力',
+		'擦傷',
+		'挫傷',
+		'扭拉傷',
+		'痠、痛',
+		'痠軟無力',
+		'僵硬',
+		'撕裂痛',
+		'活動度受限',
+		'活動時疼痛',
+		'痘痘',
+		'皮膚搔癢',
+		'皮膚紅腫',
+		'皮膚痛',
+		'皮膚麻木',
+		'水泡 / 囊泡',
+		'發疹',
+		'白斑',
+		'化膿',
+		'脫屑'
+	],
+	right_calves: [
+		'感覺異常',
+		'肢體麻痺/無力',
+		'擦傷',
+		'挫傷',
+		'扭拉傷',
+		'痠、痛',
+		'痠軟無力',
+		'僵硬',
+		'撕裂痛',
+		'活動度受限',
+		'活動時疼痛',
+		'痘痘',
+		'皮膚搔癢',
+		'皮膚紅腫',
+		'皮膚痛',
+		'皮膚麻木',
+		'水泡 / 囊泡',
+		'發疹',
+		'白斑',
+		'化膿',
+		'脫屑'
+	],
+	left_foot: [
+		'感覺異常',
+		'肢體麻痺/無力',
+		'擦傷',
+		'挫傷',
+		'扭拉傷',
+		'痠、痛',
+		'痠軟無力',
+		'僵硬',
+		'撕裂痛',
+		'活動度受限',
+		'活動時疼痛'
+	],
+	right_foot: [
+		'感覺異常',
+		'肢體麻痺/無力',
+		'擦傷',
+		'挫傷',
+		'扭拉傷',
+		'痠、痛',
+		'痠軟無力',
+		'僵硬',
+		'撕裂痛',
+		'活動度受限',
+		'活動時疼痛'
+	]
+};
